+++ safeincave/Simulation/Simulators/__init__.py
@@ -8,10 +8,14 @@ from .base import Simulator  # noqa: F401
 from .thermo_mechanical import Simulator_TM  # noqa: F401
 from .mechanical import Simulator_M  # noqa: F401
 from .thermal import Simulator_T  # noqa: F401
+from .geostatic import GeostaticStep, GeostaticReport, check_equilibrium  # noqa: F401
 
 __all__ = [
     "Simulator",
     "Simulator_TM",
     "Simulator_M",
     "Simulator_T",
+    "GeostaticStep",
+    "GeostaticReport",
+    "check_equilibrium",
 ]